requirements.txt
gunicorn
gevent
flask
pyyaml
ruamel.yaml
requests
